assert does asserts conditions (TS 3.7)

+++ packages/keyring/src/pair/encode.ts
@@ -14,7 +14,7 @@ export default function encode ({ publicKey, secretKey }: PairInfo, passphrase?:
 
   const encoded = u8aConcat(
     PKCS8_HEADER,
-    secretKey as Uint8Array,
+    secretKey,
     PKCS8_DIVIDER,
     publicKey
   );
